+++ CHANGELOG.md
@@ -7,5 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+* `hx_get`, `hx_post`, `hx_put`, `hx_patch` and `hx_delete` attributes
+* `hx_select` and `hx_target` attributes
+* `hx_swap` and `hx_swap_*` attributes
+* `hx_push_url` and `hx_push_url_str` attributes
+* `hx_swap_oob` and `hx_swap_oob_swap` attributes
+
 
 [Unreleased]: https://github.com/jcornaz/fun-htmx/compare/...HEAD